Add files via upload
These are photos that I've taken within each of the VM's that I use are to show that all of them are able to communicate with each other within a safe and controlled environment, that being my home lab. I ran the ping -c (number) <target-ip> command which basically is used to see if a system is up and able to be reached.